+++ ebird/__init__.py
@@ -7,7 +7,7 @@ from ebird.core import geo_observations, geo_notable, geo_species, \
     region_observations, region_notable, region_species, \
     hotspot_observations, hotspot_notable, hotspot_species, \
     location_observations, location_notable, location_species, \
-    nearest_species, find_locations, list_locations
+    nearest_species, find_locations, list_locations, list_hotspots
 
 VERSION = (0, 2, 0, 'final')